Create two high priority tasks to wait on lock
So that we can test if we inherit the waiter with the highest priority.

Fix a bug in get_nice() as nice == -1 is a valid return value although
the man page indicates it as an error. Check errno as well to verify we
hit a real error or just the current nice is -1.

Signed-off-by: Qais Yousef <qyousef@layalina.io>